Added SAVE_TO. Test was broken, so fixed.

+++ golang/FileList/FileList_test.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 # FileList system testing
 #   _FORCE=Y _WITH_S3=Y ./FileList_test.sh
+#
+#   source ./FileList_test.sh
+#   cd ./nxrm-3.62.0-01
+#   TestFileList 0 default
 
 # Download automate scripts
 source /dev/stdin <<<"$(curl -sfL "https://raw.githubusercontent.com/hajimeo/samples/master/bash/utils.sh" --compressed)" || return $?
@@ -12,18 +16,20 @@ _FILE_LIST="${1}" # If not specified, will try from the PATH
 _DBUSER="${2:-"nxrm"}"
 _DBPWD="${3:-"${_DBUSER}123"}"
 _DBNAME="${4:-"${_DBUSER}filelisttest"}" # Use some unique name (lowercase)
-_VER="${5:-"3.53.0-01"}"         # Specify your nexus version
+_VER="${5:-"3.61.0-02"}"         # Specify your nexus version (3.62 and 3.63 are not suitable)
 _NEXUS_TGZ_DIR="${6:-"$HOME/.nexus_executable_cache"}"
 
+#_FORCE="N"
+#_NOT_STOP_AFTER_TEST="N"
 _NEXUS_DEFAULT_PORT="18081"
 _TEST_REPO_NAME="raw-hosted"
-_ASSET_CREATE_NUM=1000
+_ASSET_CREATE_NUM=100
 _PID=""
+#_WITH_S3="N"
 _AWS_S3_BUCKET="apac-support-bucket"
 _AWS_S3_PREFIX="filelist_test"
 _S3_BLOBTORE_NAME="s3-test"
-#_FORCE="N"
-#_WITH_S3="N"
+# Currently can't specify S3 repo name, always raw-s3-hosted
 
 function TestFileList() {
     local _expected_num="${1:-0}"
@@ -34,15 +40,20 @@ function TestFileList() {
         _log "WARN" "_expected_num (${_expected_num}) might be incorrect"
     fi
 
+    # -b = base directory or S3 bucket name, -p sub directory prefix usually always 'vol-'
     local _opt="-b ./sonatype-work/nexus3/blobs/${_bsName}/content -p vol-"
     [[ "${_is_s3}" =~ ^[yY] ]] && _opt="-b ${_AWS_S3_BUCKET} -p ${_AWS_S3_PREFIX%/}/content/vol- -S3"
     if [ -s ./sonatype-work/nexus3/etc/fabric/nexus-store.properties ]; then
+        # -db = providing DB conn start checking Orphaned blobs finding because _TRUTH = 'BS'
         _opt="${_opt} -db ./sonatype-work/nexus3/etc/fabric/nexus-store.properties"
     fi
 
-    local _cmd="${_FILE_LIST} -c 10 -RF -bsName ${_bsName} ${_opt}"
+    local _cmd="${_FILE_LIST:-"file-list"} -c 10 -bsName ${_bsName} ${_opt}"
     _log "INFO" "Executing '${_cmd}' ..."
-    time ${_cmd} -X >./file-list.out 2>./file-list.log || return $?
+    if ! time ${_cmd} -X >./file-list.out 2>./file-list.log; then
+        cat ./file-list.log
+        return 1
+    fi
     local _file_list_ln="$(cat ./file-list.out | wc -l | tr -d '[:space:]')"
 
     if [ 0 -eq ${_file_list_ln} ]; then
@@ -80,24 +91,28 @@ function main() {
         fi
     fi
 
-    # Not perfect test but better than not checking...
-    if psql -l | grep -q "^ ${_DBNAME} "; then
-        _log "WARN" "Database ${_DBNAME} exists (_FORCE = ${_FORCE})"
+    _PID="$(ps auxwww | grep -F 'org.sonatype.nexus.karaf.NexusMain' | grep -vw grep | awk '{print $2}' | tail -n1)"
+    local _port=""
+    if [ -n "${_PID}" ]; then
+        _log "WARN" "NexusMain is running wit PID=${_PID} (_FORCE = ${_FORCE})"
         if [[ ! "${_FORCE}" =~ ^[yY] ]]; then
-            echo "To force, set _FORCE=\"Y\" env variable (but may not start due to Nuget repository)." >&2
-            return 1
+            echo "As no _FORCE=\"Y\", reusing this NexusMain." >&2
+            sleep 3
+            _port="${_NEXUS_DEFAULT_PORT}"  # this is not accurate ...
         else
-            _log "WARN" "As _FORCE = ${_FORCE}, will execute 'DROP DATABASE ${_DBNAME}' in 5 seconds..."
+            _log "WARN" "As _FORCE = ${_FORCE}, will execute 'kill ${_PID}\"' in 5 seconds..."
             sleep 5
-            PGPASSWORD="${_DBPWD}" psql -U ${_DBUSER} -h $(hostname -f) -p 5432 -d template1 -c "DROP DATABASE ${_DBNAME}" || return $?
+            kill ${_PID} || return $?
+            _PID=""
+            sleep 5
         fi
     fi
 
     if [ -d "./nxrm-${_VER}" ]; then
         _log "WARN" "./nxrm-${_VER} exists (_FORCE = ${_FORCE})"
         if [[ ! "${_FORCE}" =~ ^[yY] ]]; then
-            echo "To force, set _FORCE=\"Y\" env variable." >&2
-            return 1
+            echo "As no _FORCE=\"Y\", reusing this installation." >&2
+            sleep 3
         else
             _log "WARN" "As _FORCE = ${_FORCE}, will execute 'rm -rf \"./nxrm-${_VER}\"' in 5 seconds..."
             sleep 5
@@ -105,7 +120,22 @@ function main() {
         fi
     fi
 
-    mkdir -v ./nxrm-${_VER} || return $?
+    # Not perfect test but better than not checking...
+    if psql -l | grep -q "^ ${_DBNAME} "; then
+        _log "WARN" "Database ${_DBNAME} exists (_FORCE = ${_FORCE})"
+        if [[ ! "${_FORCE}" =~ ^[yY] ]]; then
+            echo "As no _FORCE=\"Y\", reusing this database (but may not start due to Nuget repository)." >&2
+            sleep 3
+        else
+            _log "WARN" "As _FORCE = ${_FORCE}, will execute 'DROP DATABASE ${_DBNAME}' in 5 seconds..."
+            sleep 5
+            PGPASSWORD="${_DBPWD}" psql -U ${_DBUSER} -h $(hostname -f) -p 5432 -d template1 -c "DROP DATABASE ${_DBNAME}" || return $?
+        fi
+    fi
+
+    if [ ! -d "./nxrm-${_VER}" ]; then
+        mkdir -v ./nxrm-${_VER} || return $?
+    fi
     cd ./nxrm-${_VER} || return $?
 
     if [ ! -s "nexus-${_VER}/bin/nexus" ]; then
@@ -127,7 +157,10 @@ function main() {
     if [ ! -s "./sonatype-work/nexus3/etc/nexus.properties" ]; then
         touch "./sonatype-work/nexus3/etc/nexus.properties"
     fi
-    local _port="$(_find_port "${_NEXUS_DEFAULT_PORT}")"
+
+    if [ -z "${_port}" ]; then
+        _port="$(_find_port "${_NEXUS_DEFAULT_PORT}")"
+    fi
     _upsert "./sonatype-work/nexus3/etc/nexus.properties" "application-port" "${_port}" || return $?
     _upsert "./sonatype-work/nexus3/etc/nexus.properties" "nexus.datastore.enabled" "true" || return $?
     _upsert "./sonatype-work/nexus3/etc/nexus.properties" "nexus.security.randompassword" "false" || return $?
@@ -145,9 +178,11 @@ advanced=maxLifetime\=600000
 EOF
 
     # Start your nexus. Not using 'start' so that stopping is easier (i think)
-    ./nexus-${_VER}/bin/nexus run &>./nexus_run.out &
-    _PID=$!
-    _log "INFO" "Executing './nexus-${_VER}/bin/nexus run &> ./nexus_run.out &' (PID: ${_PID}) ..."
+    if [ -z "${_PID}" ]; then
+        ./nexus-${_VER}/bin/nexus run &>./nexus_run.out &
+        _PID=$!
+        _log "INFO" "Executing './nexus-${_VER}/bin/nexus run &> ./nexus_run.out &' (PID: ${_PID}) ..."
+    fi
     #tail -f ./sonatype-work/nexus3/log/nexus.log
 
     # Wait until port is ready
@@ -161,46 +196,64 @@ EOF
     # always creating 'raw-hosted'
     _log "INFO" "Executing 'f_setup_raw' (creating 'raw-hosted') ..."
     f_setup_raw || return $?
-    _log "INFO" "Executing 'f_upload_dummies \"${_NEXUS_URL%/}/repository/raw-hosted/test\" \"${_ASSET_CREATE_NUM}\"' ..."
-    f_upload_dummies "${_NEXUS_URL%/}/repository/raw-hosted/test" "${_ASSET_CREATE_NUM}" || return $?
+    _log "INFO" "Executing 'f_upload_dummies_raw \"raw-hosted\" \"${_ASSET_CREATE_NUM}\"' (${_NEXUS_URL%/}) ..."
+    sleep 3
+    f_upload_dummies_raw "raw-hosted" "${_ASSET_CREATE_NUM}" || return $?
     if [[ "${_WITH_S3}" =~ ^[yY] ]]; then
-        _log "INFO" "Creating '${_S3_BLOBTORE_NAME}' (this also creates 'raw-s3-hosted' repo)..."
+        _log "INFO" "Creating S3 '${_S3_BLOBTORE_NAME}' (this also creates 'raw-s3-hosted' repo)..."
         if ! f_create_s3_blobstore "${_S3_BLOBTORE_NAME}" "${_AWS_S3_PREFIX}" "${_AWS_S3_BUCKET}"; then
             # other params use AWS_XXXX env variables
             _log "WARN" "Not testing S3 as f_create_s3_blobstore failed. Please make sure the AWS_ env variables are set."
             _WITH_S3=""
         else
-            _log "INFO" "Executing 'f_upload_dummies \"${_NEXUS_URL%/}/repository/raw-s3-hosted/test\" \"${_ASSET_CREATE_NUM}\"' ..."
-            if ! f_upload_dummies "${_NEXUS_URL%/}/repository/raw-s3-hosted/test" "${_ASSET_CREATE_NUM}"; then
-                _log "WARN" "Not testing S3 as f_upload_dummies failed. Please investigate nexus.log."
+            _log "INFO" "Executing 'f_upload_dummies_raw \"raw-s3-hosted\" \"${_ASSET_CREATE_NUM}\"' (${_NEXUS_URL%/}) ..."
+            if ! f_upload_dummies_raw "raw-s3-hosted" "${_ASSET_CREATE_NUM}"; then
+                _log "WARN" "Not testing S3 as f_upload_dummies_raw failed. Please investigate nexus.log."
                 _WITH_S3=""
             fi
         fi
     fi
 
+    _log "INFO" "Deleting all assets (${_NEXUS_URL%/}) ..."
     f_delete_all_assets "" "Y"
-    local _expected_num=$(cat /tmp/f_delete_all_assets_$$.out | wc -l | tr -d '[:space:]')
+    local _expected_num=$(cat /tmp/f_get_all_assets_$$.out | wc -l | tr -d '[:space:]')
+    sleep 3
+    _log "INFO" "Running assetBlob.cleanup (${_NEXUS_URL%/}) ..."
+    f_run_tasks_by_type "assetBlob.cleanup"
+    sleep 5
+    if f_run_tasks_by_type "" | grep '"currentState": "RUNNING"'; then
+        sleep 5
+    fi
 
     # For measuring the time of your restoring command, delete Linux (file) cache
     #sudo sh -c 'echo 3 > /proc/sys/vm/drop_caches'
 
-    _log "INFO" "Executing 'TestFileList \"${_expected_num}\" \"${_BLOBTORE_NAME}\"' ..."
-    TestFileList "${_expected_num}" "${_BLOBTORE_NAME}"
+    _log "INFO" "Executing 'TestFileList \"${_expected_num}\" \"${_BLOBTORE_NAME:-"default"}\"' ..."
+    sleep 3
+    TestFileList "${_expected_num}" "${_BLOBTORE_NAME:-"default"}"
+    if ! TestFileList "${_expected_num}" "${_BLOBTORE_NAME:-"default"}"; then
+        _log "INFO" "'TestFileList failed ($?)'"
+        return 1
+    fi
     if [[ "${_WITH_S3}" =~ ^[yY] ]]; then
         _log "INFO" "Executing 'TestFileList \"${_expected_num}\" \"${_S3_BLOBTORE_NAME}\"' ..."
-        TestFileList "${_expected_num}" "${_S3_BLOBTORE_NAME}" "${_WITH_S3}"
+        if ! TestFileList "${_expected_num}" "${_S3_BLOBTORE_NAME}" "${_WITH_S3}"; then
+            _log "INFO" "'TestFileList failed ($?)'"
+            return 1
+        fi
     fi
 
     echo "Extra manual test: "
     echo "    Start this nexus and wait or run *all* 'Cleanup unused asset blob' tasks."
+    echo "    https://sonatype.atlassian.net/browse/NEXUS-40708, so needs to wait for one hour. If version is 3.62 and 3.63"
     echo "    Re-run file-list command to make sure 'deleted=true' is removed."
-    echo "    ${_FILE_LIST} ${_opt} -c 10 -db ./sonatype-work/nexus3/etc/fabric/nexus-store.properties -RF -bsName ${_bsName} -dF $(date '+%Y-%m-%d') -RDel -X > ./file-list_del.out 2> ./file-list_del.log"
+    echo "    ${_FILE_LIST} ${_opt} -c 10 -db ./sonatype-work/nexus3/etc/fabric/nexus-store.properties -bsName ${_bsName} -dF $(date '+%Y-%m-%d') -RDel -X > ./file-list_del.out 2> ./file-list_del.log"
     echo "    Run the Reconcile with Since 1 days and with file-list.out."
 }
 
 if [ "$0" = "$BASH_SOURCE" ]; then
     main
-    if [ -n "${_PID}" ]; then
+    if [ -n "${_PID}" ] && [[ ! "${_NOT_STOP_AFTER_TEST}" =~ ^[yY] ]]; then
         _log "INFO" "Stopping Nexus (${_PID}) ..."
         sleep 3
         kill ${_PID}
